refactor post actions and reducer
