+++ packages/js/src/redux/actions/index.js
@@ -38,4 +38,4 @@ export * from "./WincherModal";
 export * from "./WincherRequest";
 export * from "./WincherSEOPerformance";
 export * from "./isPremium";
-export * from "./postId";
+export * from "./post";
